requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.13
django-allauth==0.63.3
django-bootstrap4==24.3
django-crispy-forms==1.12.0
django-environ==0.11.2
psycopg2-binary==2.9.9
soupsieve==2.5
sqlparse==0.5.0
